% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kdict_formants.R
\name{read_kdict_formants}
\alias{read_kdict_formants}
\title{Read a tier from all TextGrids within the kdict root directory}
\usage{
read_kdict_formants(kdict_path, tracker = c("praat", "forest"),
  .progress = FALSE)
}
\arguments{
\item{kdict_path}{path to where all kdict batch folder are (e.g. b001, b002, etc.)}

\item{tracker}{name(s) of formant tracker ('praat' and/or 'forest'; default is both)}

\item{.progress}{print out progress bar during read? (default is \code{FALSE})}
}
\description{
Read a tier from all TextGrids within the kdict root directory
}
