% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kdict_tiers.R
\name{read_kdict_tiers}
\alias{read_kdict_tiers}
\title{Read a tier from all TextGrids within the kdict root directory}
\usage{
read_kdict_tiers(kdict_path, tier_names, .progress = FALSE)
}
\arguments{
\item{kdict_path}{path to where all kdict batch folder are (e.g. b001, b002, etc.)}

\item{.progress}{print out progress bar during read? (default is \code{FALSE})}

\item{tier_name}{name(s) of TextGrid tier ('ipa', 'vowels', 'formants', 'tags', or 'comments')}
}
\description{
Read a tier from all TextGrids within the kdict root directory
}
