% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_vowels.R
\name{wrangle_vowels}
\alias{wrangle_vowels}
\alias{read_vowel_defs}
\alias{remove_diacritics}
\alias{get_cons_context}
\title{Helper functions for wrangling kdict vowel- and word-level transcriptions}
\usage{
read_vowel_defs()

remove_diacritics(ipa_string)

get_cons_context(transcription, vowels)
}
\arguments{
\item{ipa_string}{a string or a vector of strings from which all diacritics are to be removed}

\item{transcription}{kdict transcription from the 'ipa' tier, e.g. \code{'aleke'}}

\item{vowels}{a vector of kdict vowels from the 'vowel' tier, e.g. \code{c('a', 'e', 'e')}}
}
\description{
Helper functions for wrangling kdict vowel- and word-level transcriptions
}
\section{Wrangling functions}{


\describe{
   \item{`get_cons_context()`}{returns a vector of consonantal contexts for the
   vowels within a given transcription. Will return an error if vowels vector and
   transcription do not have identical vowels.}

   \item{`read_vowel_defs()`}{returns the table of vowel definitions and features
   contained within the package.}

   \item{`remove_diacritics()`}{removes all IPA diacritics input string.}
}
}

\examples{
get_cons_context('ŋɐeelə', c('ɐe', 'e', 'ə')) # returns: c("ŋ_V", "V_l", "l_#")

}
